requirements.txt
fastapi==0.109.2
uvicorn
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0